Uploading C++ json reader class and blank .json
The C++ class needs the json.hpp file, which can be found at the repository https://github.com/nlohmann/json. Also, the unrealConfig.json file is currently empty and will be for storing the redis config settings as they come in. configReader.cpp is intended to read unrealConfig.json. Code review will be very necessary for the C++ file.